Added support for .creator_profile (needs testing!)
The current implementation simply uses the file as a unit
and the unit identifier is constructed from the filename
and the prefix "static|". I'm not entirely sure if this is
safe since it could break when the identifier is validated
at a point where the "static|" prefix is not checked.

Corresponds to issue #19.